FIX - Check if directory is a symlink when performing candidate upgrade.

+++ src/main/bash/selfupdate.sh
@@ -189,7 +189,7 @@ SDKMAN_CANDIDATES=(${SDKMAN_CANDIDATES_CSV})
 IFS="$OLD_IFS"
 
 for candidate in "${SDKMAN_CANDIDATES[@]}"; do
-    if [[ -n "$candidate" && -d "${SDKMAN_DIR}/${candidate}" ]]; then
+    if [[ -n "$candidate" && -d "${SDKMAN_DIR}/${candidate}" && ! -L "${SDKMAN_DIR}/${candidate}" ]]; then
         if [[ -z "$(ls -A ${SDKMAN_DIR}/${candidate})" ]]; then
             sdkman_echo_debug "Attempt removal of ${candidate} dir: ${SDKMAN_DIR}/${candidate}"
             rmdir "${SDKMAN_DIR}/${candidate}"
